requirements.txt
flask==2.3.3
pillow==10.0.1
numpy==1.24.3
tensorflow==2.13.0